context-engineering-workflow/pyproject.toml
[project]
name = "context-engineering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=0.165.1",
    "crewai-tools>=0.0.1",
    "ipykernel>=6.30.1",
    "openai>=1.99.9",
    "pymilvus>=2.6.0",
    "tensorlake>=0.2.44",
    "voyageai>=0.3.4",
    "zep-crewai>=1.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "firecrawl-py>=4.3.0",
    "streamlit>=1.49.1",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
